+++ src/App/common/MovieTile/index.js
@@ -13,7 +13,7 @@ import {
   PosterIMG,
 } from "./styled";
 import { genres } from "../genreIds";
-import { IMG_URL } from "../../common/ApiUrls";
+import { IMG_URL } from "../Helpers/ApiUrls";
 import whenNoPoster from "../../images/whenNoPoster.png";
 
 const MovieTile = ({ posterPath, title, year, tag, rating, votes }) => {
