+++ src/App/features/Details/MovieDetails/MovieInfo/styled.js
