+++ src/App/features/Details/MovieDetails/MovieTop/styled.js
