+++ src/App/features/Details/PersonDetails/index.js
@@ -1,12 +1,15 @@
-import { Main } from "../../common/Main";
+import { Main } from "../../../common/Main";
 import About from "./About";
 import { ProfileWrapper } from "./styled";
-import Error from "../../common/Error";
-import Loading from "../../common/Loading";
-import MoviesList from "../../common/MoviesList";
-import MainHeader from "../../common/MainHeader";
+import Error from "../../../common/Error";
+import Loading from "../../../common/Loading";
+import MoviesList from "../../../common/MoviesList";
+import MainHeader from "../../../common/MainHeader";
 import { useParams } from "react-router-dom/cjs/react-router-dom.min";
-import { fetchPersonDataById, selectPersonData } from "../Details/detailsSlice";
+import {
+  fetchPersonDataById,
+  selectPersonData,
+} from "../../Details/detailsSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from "react";
