+++ src/App/features/Details/PersonDetails/styled.js
